requirements.txt
Django==2.0.4
dj-database-url==0.5.0
django-bootstrap4==0.0.6
django-extensions==2.0.7
django-test-plus==1.0.22
python-decouple==3.1
django-test-without-migrations==0.6
pytz
